Rust + Aync + TCP sockets = Fun!
The TCP server and associated service (found within the `service` crate) reads
in requests and constructs responses in an asynchronous fashion. Requests are
serialized into a Message (with the use of the zerocopy crate), the message's
header and payload are validated then, the appropriate response is constructed
and sent to the client.